Update SMADB_Backup.sql
Via WITH INIT the BACKUP DATABASE in this script overwrites the existing database backup with a new one. The BACKUP LOG should do the same. When the script runs you back up the database as that point in time. There is no need to retain preceding transaction log records, therefore, BACKUP LOG should also be WITH INIT. Thank you.